# backend/public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'Quantigo API Documentation'
  description: 'Quantigo is a commerce coordination, procurement, merchant operating system, and pooled purchasing platform.'
  version: 1.0.0
servers:
  - url: 'http://localhost:8568'
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'You can retrieve your token by visiting the login endpoint.'
security:
  -
    default: []
paths:
...
  /api/v1/auth/reset-password:
    post:
      summary: 'Reset password'
      operationId: resetPassword
      description: 'Update the account password using the token sent in the reset link.'
      parameters: []
      responses: {}
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                  description: ''
                  example: architecto
                email:
                  type: string
                  description: 'Must be a valid email address. Must match an existing stored value.'
                  example: zbailey@example.net
                password:
                  type: string
                  description: 'Must be at least 8 characters.'
                  example: '-0pBNvYgxw'
              required:
                - token
                - email
                - password
      security: []
...
  /api/v1/auth/kyb/check-hash:
    post:
      summary: 'Check asset hash'
      operationId: checkAssetHash
      description: 'Check if a file content hash already exists in our local registry to bypass uploading again.'
      parameters: []
      responses: {}
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                hash:
                  type: string
                  description: 'Must be 64 characters.'
                  example: bngzmiyvdljnikhwaykcmyuwpwlvqwrsitcpscqldzsnrwtujwvlxjklqppwqbew
              required:
                - hash
...